Attempted to implement accountRoutes with editPassword
